Adapting flutter package best practices
