--- flutter-vapi.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "../Downloads/daily_flutter-0.1.0-alpha.4-1"
-		}
-	],
-	"settings": {}
-}
