+++ lib/src/types/types.dart
@@ -1,0 +1,13 @@
+/// Vapi Flutter SDK Type Definitions
+/// 
+/// This library exports all the type definitions used by the Vapi Flutter SDK,
+/// including events, exceptions, and enums.
+
+// Export all exception classes
+export 'errors.dart';
+
+// Export event types
+export 'vapi_event.dart';
+
+// Export enums and other types
+export 'vapi_audio_device.dart';
